% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myKmeans.R
\name{Center_change}
\alias{Center_change}
\title{Center Change}
\usage{
Center_change(center_former, center_present, tol)
}
\arguments{
\item{center_former}{a matrix, cluster centers before iteration.}

\item{center_present}{a matrix, cluster centers after iteration.}

\item{tol}{numeric, tolerance value of center change.}
}
\value{
A bool. If the difference of the centers before iteration and the centers after iteration is smaller
than the tolerance number, the returned value is \code{TRUE}, otherwise \code{FALSE}.
}
\description{
This function verifies if the cluster centers change after each iteration.
}
\examples{
data <- iris[,3:4]
center_former <- data[c(1,3,4),]
center_present <- data[c(1,3,5),]
tol <- 1e-6
center_change = Center_change(center_former,center_present,tol)
center_change
}
