% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myKmeans.R
\name{Center_init}
\alias{Center_init}
\title{Cluster Center Initiation}
\usage{
Center_init(data, k)
}
\arguments{
\item{data}{a matrix, each row represents a sample.}

\item{k}{numeric, number of cluster centers.}
}
\value{
A matrix consisting of \code{k} randomly chosen initial cluster
centers. Each row represents a cluster center.
}
\description{
This function randomly chooses \code{k} initial cluster centers of the K-means
algorithm.
}
\examples{
data <- iris[,3:4]
Center_init(data,3)
}
\seealso{
See \code{\link{myKmeans}} for more information about inputs.
}
