% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myKmeans.R
\name{cluster_center}
\alias{cluster_center}
\title{Cluster Center Update}
\usage{
cluster_center(data, cluster_affiliation, k)
}
\arguments{
\item{data}{a matrix, each row represents a sample.}

\item{cluster_affiliation}{a vector, each element represents the cluster affiliation of each sample.}

\item{k}{numeric, number of cluster centers.}
}
\value{
A matrix. It has the same form and dimension as the cluster centers before iteration.
}
\description{
This function updates the cluster centers after iteration.
}
\details{
This function computes the new cluster centers by averaging the coordinates of
the samples affiliated to each cluster.
}
\examples{
data <- iris[,3:4]
cluster_affiliation <- sample(c(1,2,3), nrow(data), replace = TRUE)
center_new <- cluster_center(data,cluster_affiliation,3)
}
\seealso{
See \code{\link{Nearest_cluster}} for more information about the
computation of the cluster affiliation of a sample.
}
