% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myKmeans.R
\name{cluster_visual}
\alias{cluster_visual}
\title{Cluster Visualization}
\usage{
cluster_visual(data, var_names, kmeans_result, cluster_origin = NA)
}
\arguments{
\item{data}{a matrix, each row represents a sample. It should be the same as
what was used in the K-means algorithm.}

\item{var_names}{a (two-dimensional) vector. The names of the coordinates used
in the K-means algorithm.}

\item{kmeans_result}{a list, output of the K-means algorithm.}

\item{cluster_origin}{a vector, the original cluster affiliations of samples (if accessible). Default is NA.}
}
\value{
A plot of the clustering result of the K-means algorithm and an
(optional) plot of the original cluster affiliations of samples.
}
\description{
This function visualizes the clustering results of the K-means algorithm.
}
\details{
This function makes use of \code{ggplot2} and plots the clustering results of
the K-means algorithm. Users can also enters the original clusters for
comparison. It currently only supports two-dimensional visualization.
}
\examples{
data <- iris[,c(3:4)]
mk <- myKmeans(data, k = 3)
cv <- cluster_visual(data,colnames(iris)[3:4],mk)
cv
}
\seealso{
See \code{\link{myKmeans}} for more information about the output of the
K-means algorithm.
}
