% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myKmeans.R
\name{Dist}
\alias{Dist}
\title{Euclidean Distance}
\usage{
Dist(a, b)
}
\arguments{
\item{a, b}{numeric vectors, the euclidean distance is computed as the square
root of the summation of the square of the difference between each
coordinate of \code{a} and \code{b}. \code{a} and \code{b} should have the
same length.}
}
\value{
A numeric value representing the euclidean distance between point
\code{a} and \code{b}.
}
\description{
This function computes the euclidean distance of two sample points.
}
\examples{
a <- c(1,7,4)
b <- c(2,0,5)
Dist(a,b)
}
