% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iris.R
\docType{data}
\name{iris}
\alias{iris}
\title{Characters and Categories of 150 iris data}
\format{
A data frame with 150 rows and 5 variables:
\describe{
\item{Sepal.Length}{Sepal length, in cm}
\item{Sepal.Width}{Sepal width, in cm}
\item{Petal.Length}{Petal length, in cm}
\item{Petal.Width}{Petal width, in cm}
\item{Species}{factor, categories of iris}
}
}
\source{
\url{http://archive.ics.uci.edu/ml/datasets/Iris}
}
\usage{
iris
}
\description{
A dataset containing the length and width of the sepal and petal and
categories of 150 iris.
}
\keyword{datasets}
