% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myKmeans.R
\name{Kmeans_check}
\alias{Kmeans_check}
\title{Input Check}
\usage{
Kmeans_check(data, centers, k)
}
\arguments{
\item{data}{a matrix, each row represents a sample.}

\item{centers}{a matrix, each row represents a cluster center.}

\item{k}{numeric, number of cluster centers.}
}
\description{
This function checks the inputs of the K-means algorithm.
}
\details{
This function checks if users enter datasets and enter the matrix of cluster
centers or valid number of cluster centers.
}
\examples{
data <- iris[,3:4]
center <- data[c(1,3,4),]

Kmeans_check(data,center,NA)

Kmeans_check(data,NA,k=3)

\dontrun{
Kmeans_check(data,NA,0)
}
}
\seealso{
See \code{\link{myKmeans}} for more information about inputs.
}
