% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myKmeans.R
\name{myKmeans}
\alias{myKmeans}
\alias{kmeans}
\title{K-means Algorithm}
\usage{
myKmeans(data = NA, centers = NA, k = NA)
}
\arguments{
\item{data}{a matrix, each row represents a sample.}

\item{centers}{a matrix, each row represents a cluster center. Default is NA.}

\item{k}{numeric, number of cluster centers. Default is NA.}
}
\value{
An object of class "\code{list}" which has two components. The first
component is a vector of final cluster affiliations of samples. The second
component is a matrix consisting of the final cluster centers.
}
\description{
This function implements the K-means algorithm.
}
\details{
The K-means algorithm was first proposed by James MacQueen (1967)
\url{Doi:10.1.1.308.8619}. Users can either specify the initial cluster centers
directly or just the number of cluster centers.
}
\examples{
data <- iris[,3:4]
centers <- data[c(1,3,4),]
mk1 <- myKmeans(data, centers)

mk2 <- myKmeans(data, k = 3)
}
\seealso{
See \code{\link{Kmeans_check}} for more information about input check.
}
