% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{myKmeans_C}
\alias{myKmeans_C}
\title{K-means Algorithm Using C++}
\usage{
myKmeans_C(data, centers = matrix(2, 2), k = 0L)
}
\arguments{
\item{data}{a matrix, each row represents a sample.}

\item{centers}{a matrix, each row represents a cluster center. Default is NA.}

\item{k}{numeric, number of cluster centers. Default is NA.}
}
\description{
This function implements the K-means algorithm using C++.
}
\details{
The K-means algorithm was first proposed by James MacQueen (1967)
\url{Doi:10.1.1.308.8619}. Users can either specify the initial cluster centers
directly or just the number of cluster centers.
}
