% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myKmeans.R
\name{Nearest_cluster}
\alias{Nearest_cluster}
\title{Nearest Cluster Center}
\usage{
Nearest_cluster(a, cluster_centers, k)
}
\arguments{
\item{a}{a numeric vector which represents a sample.}

\item{cluster_centers}{a matrix, each row represents a cluster center.}

\item{k}{numeric, number of cluster centers.}
}
\value{
A numeric value. Row subscript of the nearest cluster center of sample
\code{a} in cluster center matrix.
}
\description{
This function finds the nearest cluster center of a sample.
}
\details{
This function first computes the distance between a sample and each cluster
center, and finds the most close one to the sample, i.e. specify the cluster
affiliation of the sample.
}
\examples{
data <- iris[,3:4]
cluster_centers <- data[c(1,3,4),]
a <- data[14,]
Nearest_cluster(a,cluster_centers,3)
}
